Separating StandardApiAction to StandardApiAction and StandardApiActionImpl
