+++ src/main/java/org/ideoholic/curium/model/std/action/StandardApiAction.java
@@ -4,17 +4,13 @@
 package org.ideoholic.curium.model.std.action;
 
 import org.ideoholic.curium.dto.ResultResponse;
-import org.ideoholic.curium.exceptions.CustomErrorMessage;
-import org.ideoholic.curium.exceptions.CustomResponseException;
 import org.ideoholic.curium.model.std.dto.ClassDto;
 import org.ideoholic.curium.model.std.dto.ClassIdsDto;
 import org.ideoholic.curium.model.std.dto.ClassesHierarchyDto;
 import org.ideoholic.curium.model.std.dto.StdOfClassDto;
 import org.ideoholic.curium.model.std.dto.UpperLowerClassDto;
-import org.ideoholic.curium.model.std.service.StandardService;
 import org.ideoholic.curium.model.student.dto.StudentIdsDto;
 import org.ideoholic.curium.util.Constants;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -22,141 +18,65 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
 
 
-@RestController
+
 @RequestMapping("/api/v1/classProcess")
-public class StandardApiAction {
-
-    @Autowired
-    private StandardService standardService;
-
+public interface StandardApiAction {
 
     @PostMapping("/restoreMultipleLeftout")
-    public ResponseEntity<ResultResponse> restoreMultipleLeftout(@RequestBody StudentIdsDto dto) {
-        standardService.restoreMultipleLeftout(dto);
-        return viewLeftOut();
-    }
+    public ResponseEntity<ResultResponse> restoreMultipleLeftout(@RequestBody StudentIdsDto dto);
 
     @GetMapping("/viewLeftOut")
-    public ResponseEntity<ResultResponse> viewLeftOut() {
-        ResultResponse result = standardService.viewleft();
-        return ResponseEntity.ok(result);
-    }
+    public ResponseEntity<ResultResponse> viewLeftOut();
 
     @PostMapping("/searchByClass")
-    public ResponseEntity<ResultResponse> searchByClass(@RequestBody StdOfClassDto dto, @RequestHeader(value = Constants.BRANCHID) String branchId, @RequestHeader(value = Constants.CURRENTACADEMICYEAR) String currentAcademicYear) {
-        ResultResponse result = standardService.searchByClass(dto, branchId, currentAcademicYear);
-        return ResponseEntity.ok(result);
-    }
-
+    public ResponseEntity<ResultResponse> searchByClass(@RequestBody StdOfClassDto dto, @RequestHeader(value = Constants.BRANCHID) String branchId, @RequestHeader(value = Constants.CURRENTACADEMICYEAR) String currentAcademicYear);
+   
     @PostMapping("/restoreMultipleDroppedout")
-    public ResponseEntity<ResultResponse> restoreMultipleDroppedout(@RequestBody StudentIdsDto dto) {
-        standardService.restoreMultipleDroppedout(dto);
-        return viewDropped();
-    }
+    public ResponseEntity<ResultResponse> restoreMultipleDroppedout(@RequestBody StudentIdsDto dto);
 
     @PostMapping("/restoreMultipleGraduate")
-    public ResponseEntity<ResultResponse> restoreMultipleGraduate(@RequestBody StudentIdsDto dto) {
-        standardService.restoreMultipleGraduate(dto);
-        return viewGraduated();
-    }
+    public ResponseEntity<ResultResponse> restoreMultipleGraduate(@RequestBody StudentIdsDto dto);
 
     @GetMapping("/viewDropped")
-    public ResponseEntity<ResultResponse> viewDropped() {
-        ResultResponse result = standardService.viewDropped();
-        return ResponseEntity.ok(result);
-    }
+    public ResponseEntity<ResultResponse> viewDropped();
 
     @GetMapping("/viewGraduated")
-    public ResponseEntity<ResultResponse> viewGraduated() {
-        ResultResponse result = standardService.viewGraduated();
-        return ResponseEntity.ok(result);
-    }
+    public ResponseEntity<ResultResponse> viewGraduated();
 
     @PostMapping("/leftoutMultiple")
-    public ResponseEntity<ResultResponse> leftoutMultiple(@RequestBody StudentIdsDto dto) {
-        ResultResponse result = standardService.leftoutMultiple(dto);
-        if(result.isSuccess()) {
-            return ResponseEntity.ok(result);
-        }
-        throw new CustomResponseException(CustomErrorMessage.FAILURELEFTOUT);
-    }
+    public ResponseEntity<ResultResponse> leftoutMultiple(@RequestBody StudentIdsDto dto);
 
     @PostMapping("/dropoutMultiple")
-    public ResponseEntity<ResultResponse> droppedoutMultiple(@RequestBody StudentIdsDto dto) {
-        ResultResponse result = standardService.droppedoutMultiple(dto);
-        if(result.isSuccess()) {
-            return ResponseEntity.ok(result);
-        }
-        throw new CustomResponseException(CustomErrorMessage.FAILUREDROPPEDOUT);
-    }
+    public ResponseEntity<ResultResponse> droppedoutMultiple(@RequestBody StudentIdsDto dto);
 
     @PostMapping("/graduateMultiple")
-    public ResponseEntity<ResultResponse> graduateMultiple(@RequestBody StudentIdsDto dto) {
-        ResultResponse result = standardService.graduateMultiple(dto);
-        if(result.isSuccess()) {
-            return ResponseEntity.ok(result);
-        }
-        throw new CustomResponseException(CustomErrorMessage.FAILUREGRADUATE);
-    }
+    public ResponseEntity<ResultResponse> graduateMultiple(@RequestBody StudentIdsDto dto);
 
 
     @PostMapping("/deleteClassHierarchy")
-    public ResponseEntity<ClassesHierarchyDto> deleteClassHierarchy(@RequestBody ClassIdsDto dto, @RequestHeader(value = Constants.BRANCHID) String branchId) {
-        standardService.deleteClassHierarchy(dto, branchId);
-        return classHierarchy(branchId);
-    }
+    public ResponseEntity<ClassesHierarchyDto> deleteClassHierarchy(@RequestBody ClassIdsDto dto, @RequestHeader(value = Constants.BRANCHID) String branchId);
+       
 
     @PostMapping("/addClassHierarchy")
-    public ResponseEntity<ClassesHierarchyDto> addClassHierarchy(@RequestBody UpperLowerClassDto dto, @RequestHeader(value = Constants.BRANCHID) String branchId, @RequestHeader(value = Constants.USERID) String userId) {
-        standardService.addClassHierarchy(dto, branchId, userId);
-        return classHierarchy(branchId);
-    }
-
+    public ResponseEntity<ClassesHierarchyDto> addClassHierarchy(@RequestBody UpperLowerClassDto dto, @RequestHeader(value = Constants.BRANCHID) String branchId, @RequestHeader(value = Constants.USERID) String userId);
+        
     @GetMapping("/classHierarchy")
-    public ResponseEntity<ClassesHierarchyDto> classHierarchy(@RequestHeader(value = Constants.BRANCHID) String branchId) {
-        return ResponseEntity.ok(ClassesHierarchyDto.builder()
-            .classsecList(standardService.viewClasses(branchId).getResultList())
-            .classHierarchy(standardService.viewClassHierarchy(branchId).getResultList())
-            .build());
-    }
+    public ResponseEntity<ClassesHierarchyDto> classHierarchy(@RequestHeader(value = Constants.BRANCHID) String branchId);
 
     @GetMapping("/promoteClass")
-    public ResponseEntity<ResultResponse> promoteClass(@RequestHeader(value = Constants.BRANCHID) String branchId) {
-        ResultResponse result = standardService.viewClasses(branchId);
-        if(result.isSuccess()) {
-            return ResponseEntity.ok(result);
-        }
-        throw new CustomResponseException(CustomErrorMessage.ERROR);
-    }
+    public ResponseEntity<ResultResponse> promoteClass(@RequestHeader(value = Constants.BRANCHID) String branchId);
 
     @PostMapping("/deleteClass")
-    public ResponseEntity<ResultResponse> deleteClass(@RequestBody ClassIdsDto dto, @RequestHeader(value = Constants.BRANCHID) String branchId) {
-        ResultResponse result = standardService.deleteClasses(dto, branchId);
-        if(result.isSuccess()) {
-            return ResponseEntity.ok(result);
-        }
-        throw new CustomResponseException(CustomErrorMessage.ERROR);
-    }
+    public ResponseEntity<ResultResponse> deleteClass(@RequestBody ClassIdsDto dto, @RequestHeader(value = Constants.BRANCHID) String branchId);
+       
 
     @GetMapping("/viewClasses")
-    public ResponseEntity<ResultResponse> viewClasses(@RequestParam(value = Constants.BRANCHID) String branchId) {
-        ResultResponse result = standardService.viewClasses(branchId);
-        if(result.isSuccess()) {
-            return ResponseEntity.ok(result);
-        }
-        throw new CustomResponseException(CustomErrorMessage.ERROR);
-    }
-
+    public ResponseEntity<ResultResponse> viewClasses(@RequestParam(value = Constants.BRANCHID) String branchId);
+        
     @PostMapping("/createClass")
-    public ResponseEntity<ResultResponse> createClass(@RequestBody ClassDto classDto, @RequestHeader(value = Constants.BRANCHID) String branchId, @RequestHeader(value = Constants.USERID) String userId) {
-        ResultResponse result = standardService.createClass(classDto, branchId, userId);
-        if(result.isSuccess()) {
-            return ResponseEntity.ok(result);
-        }
-        throw new CustomResponseException(CustomErrorMessage.ERROR);
-    }
+    public ResponseEntity<ResultResponse> createClass(@RequestBody ClassDto classDto, @RequestHeader(value = Constants.BRANCHID) String branchId, @RequestHeader(value = Constants.USERID) String userId);
+       
 
 }
